create data, di, screen, domain layer in search_screen feature create search_screen_module.kt create and fill fragment_news_search.xml
