+++ app/src/main/java/com/example/newsfetcher/App.kt
@@ -3,9 +3,10 @@ package com.example.newsfetcher
 import android.app.Application
 import com.example.newsfetcher.di.databaseModule
 import com.example.newsfetcher.di.networkModule
-import com.example.newsfetcher.feature.bookmarks.di.bookmarksModule
-import com.example.newsfetcher.feature.detailscreen.di.detailModule
+import com.example.newsfetcher.feature.bookmarks_screen.di.bookmarksModule
+import com.example.newsfetcher.feature.details_creen.di.detailModule
 import com.example.newsfetcher.feature.main_screen.di.mainScreenModule
+import com.example.newsfetcher.feature.search_screen.di.searchScreenModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -18,7 +19,14 @@ class App : Application() {
         startKoin {
             androidLogger()
             androidContext(this@App)
-            modules(networkModule, mainScreenModule, bookmarksModule, databaseModule, detailModule)
+            modules(
+                networkModule,
+                mainScreenModule,
+                bookmarksModule,
+                databaseModule,
+                detailModule,
+                searchScreenModule
+            )
         }
 
         // Для принудительного использования темной темы приложением
