--- app/src/main/java/com/example/newsfetcher/feature/bookmarks/data/model/BookmarksRepositoryImpl.kt
@@ -1,29 +1,0 @@
-package com.example.newsfetcher.feature.bookmarks.data.model
-
-import com.example.newsfetcher.feature.bookmarks.data.BookmarksLocalSource
-import com.example.newsfetcher.feature.bookmarks.data.BookmarksRepository
-import com.example.newsfetcher.feature.bookmarks.data.toDomain
-import com.example.newsfetcher.feature.bookmarks.data.toEntity
-import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
-
-
-class BookmarksRepositoryImpl(private val bookmarksLocalSource: BookmarksLocalSource) :
-    BookmarksRepository {
-
-    override suspend fun create(model: ArticleModel) {
-        bookmarksLocalSource.create(model.toEntity())
-    }
-
-    override suspend fun read(): List<ArticleModel> {
-        return bookmarksLocalSource.read().map { it.toDomain() }
-    }
-
-    override suspend fun update(model: ArticleModel) {
-        bookmarksLocalSource.update(model.toEntity())
-    }
-
-    override suspend fun delete(model: ArticleModel) {
-        bookmarksLocalSource.delete(model.toEntity())
-    }
-
-}
