+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks_screen/data/BookmarksRepository.kt
@@ -1,0 +1,16 @@
+package com.example.newsfetcher.feature.bookmarks_screen.data
+
+import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
+
+
+interface BookmarksRepository {
+
+    suspend fun create(model: ArticleModel)
+
+    suspend fun read(): List<ArticleModel>
+
+    suspend fun update(model: ArticleModel)
+
+    suspend fun delete(model: ArticleModel)
+
+}
