+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks_screen/data/model/BookmakEntity.kt
@@ -1,0 +1,28 @@
+package com.example.newsfetcher.feature.bookmarks_screen.data.model
+
+import androidx.room.ColumnInfo
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+import com.example.newsfetcher.feature.bookmarks_screen.di.BOOKMARKS_TABLE
+
+
+@Entity(tableName = BOOKMARKS_TABLE)
+data class BookmarkEntity(
+    @PrimaryKey
+    @ColumnInfo(name = "name")
+    val name: String,
+    @ColumnInfo(name = "url")
+    val url: String,
+    @ColumnInfo(name = "title")
+    val title: String,
+    @ColumnInfo(name = "description")
+    val description: String,
+    @ColumnInfo(name = "author")
+    val author: String,
+    @ColumnInfo(name = "urlToImage")
+    val urlToImage: String,
+    @ColumnInfo(name = "publishedAt")
+    val publishedAt: String,
+    @ColumnInfo(name = "content")
+    val content: String
+)
