+++ app/src/main/java/com/example/newsfetcher/feature/bookmarks_screen/data/model/BookmarksDao.kt
@@ -1,0 +1,22 @@
+package com.example.newsfetcher.feature.bookmarks_screen.data.model
+
+import androidx.room.*
+import com.example.newsfetcher.feature.bookmarks_screen.di.BOOKMARKS_TABLE
+
+
+@Dao
+interface BookmarksDao {
+
+    @Insert(onConflict = OnConflictStrategy.REPLACE)
+    suspend fun create(entity: BookmarkEntity)
+
+    @Query("SELECT*FROM $BOOKMARKS_TABLE")
+    suspend fun read(): List<BookmarkEntity>
+
+    @Update
+    suspend fun update(entity: BookmarkEntity)
+
+    @Delete
+    suspend fun delete(entity: BookmarkEntity)
+
+}
