--- app/src/main/java/com/example/newsfetcher/feature/detailscreen/data/Mapper.kt
@@ -1,28 +1,0 @@
-package com.example.newsfetcher.feature.detailscreen.data
-
-import com.example.newsfetcher.feature.bookmarks.data.model.BookmarkEntity
-import com.example.newsfetcher.feature.detailscreen.data.model.DetailEntity
-import com.example.newsfetcher.feature.main_screen.domian.ArticleModel
-
-
-fun DetailEntity.toDomain() = ArticleModel(
-    name = name,
-    author = author,
-    title = title,
-    description = description,
-    url = url,
-    urlToImage = urlToImage,
-    publishedAt = publishedAt,
-    content = content
-)
-
-fun ArticleModel.toEntity() = DetailEntity(
-    name = name,
-    author = author,
-    title = title,
-    description = description,
-    url = url,
-    urlToImage = urlToImage,
-    publishedAt = publishedAt,
-    content = content
-)
