--- app/src/main/java/com/example/newsfetcher/feature/detailscreen/data/model/DetailEntity.kt
@@ -1,28 +1,0 @@
-package com.example.newsfetcher.feature.detailscreen.data.model
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import com.example.newsfetcher.feature.detailscreen.di.DETAIL_TABLE
-
-
-@Entity(tableName = DETAIL_TABLE)
-data class DetailEntity(
-    @PrimaryKey
-    @ColumnInfo(name = "name")
-    val name: String,
-    @ColumnInfo(name = "url")
-    val url: String,
-    @ColumnInfo(name = "title")
-    val title: String,
-    @ColumnInfo(name = "description")
-    val description: String,
-    @ColumnInfo(name = "author")
-    val author: String,
-    @ColumnInfo(name = "urlToImage")
-    val urlToImage: String,
-    @ColumnInfo(name = "publishedAt")
-    val publishedAt: String,
-    @ColumnInfo(name = "content")
-    val content: String
-)
