+++ app/src/main/java/com/example/newsfetcher/feature/main_screen/di/main_screen_module.kt
@@ -1,6 +1,5 @@
 package com.example.newsfetcher.feature.main_screen.di
 
-import com.example.newsfetcher.feature.detailscreen.di.detailModule
 import com.example.newsfetcher.feature.main_screen.data.ArticlesRemoteSource
 import com.example.newsfetcher.feature.main_screen.data.ArticlesRepository
 import com.example.newsfetcher.feature.main_screen.data.NewsAPI
