+++ app/src/main/java/com/example/newsfetcher/feature/search_screen/presentation/NewsSearchFragment.kt
@@ -1,0 +1,26 @@
+package com.example.newsfetcher.feature.search_screen.presentation
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.newsfetcher.R
+
+
+class NewsSearchFragment : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_news_search, container, false)
+    }
+
+}
